qdrant/requirements.txt
pillow
transformers
torch
qdrant-client
